all the extras for starting project
