+++ src/client/styles/header.css
@@ -1,0 +1,5 @@
+header {
+    display: flex;
+    justify-content: space-between;
+    padding: 10px 40px;
+}
