+++ src/server/mockAPI.js
@@ -1,0 +1,7 @@
+let json = {
+    'title': 'test json response',
+    'message': 'this is a message',
+    'time': 'now'
+}
+
+module.exports = json
